Even more type safety

+++ src/battle_logic.jl
@@ -2,40 +2,40 @@ using Distributions, Setfield, Match, StaticArrays
 
 function get_possible_decisions(state::IndividualBattleState; allow_nothing = false)
     @inbounds activeMon = state.teams[state.agent].mons[state.teams[state.agent].active]
-    @inbounds return @SVector [((allow_nothing || activeMon.fastMoveCooldown > 0) && activeMon.hp > 0) ? 1.0 : 0.0,
-                                ((allow_nothing || activeMon.fastMoveCooldown > 0) && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeMon.fastMoveCooldown <= 0 && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeMon.fastMoveCooldown <= 0 && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeMon.energy >= activeMon.chargedMoves[1].energy && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeMon.energy >= activeMon.chargedMoves[1].energy && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeMon.energy >= activeMon.chargedMoves[2].energy && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeMon.energy >= activeMon.chargedMoves[2].energy && activeMon.hp > 0) ? 1.0 : 0.0,
+    @inbounds return @SVector [((allow_nothing || activeMon.fastMoveCooldown > 0) && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                ((allow_nothing || activeMon.fastMoveCooldown > 0) && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.fastMoveCooldown <= 0 && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.fastMoveCooldown <= 0 && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.energy >= activeMon.chargedMoves[1].energy && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.energy >= activeMon.chargedMoves[1].energy && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.energy >= activeMon.chargedMoves[2].energy && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.energy >= activeMon.chargedMoves[2].energy && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
                                 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
 end
 
 function get_possible_decisions(state::State; allow_nothing = false)
     @inbounds activeTeam = state.teams[state.agent]
     @inbounds activeMon = activeTeam.mons[activeTeam.active]
-    @inbounds return @SVector [((allow_nothing || activeMon.fastMoveCooldown > 0) && activeMon.hp > 0) ? 1.0 : 0.0,
-                                ((allow_nothing || activeMon.fastMoveCooldown > 0) && activeMon.hp > 0) ? 1.0 : 0.0,
+    @inbounds return @SVector [((allow_nothing || activeMon.fastMoveCooldown > 0) && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                ((allow_nothing || activeMon.fastMoveCooldown > 0) && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
                                 (activeMon.fastMoveCooldown <= 0 && activeMon.hp > 0) ? 1.0 : 0.0,
                                 (activeMon.fastMoveCooldown <= 0 && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeMon.energy >= activeMon.chargedMoves[1].energy && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeMon.energy >= activeMon.chargedMoves[1].energy && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeMon.energy >= activeMon.chargedMoves[2].energy && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeMon.energy >= activeMon.chargedMoves[2].energy && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeTeam.switchCooldown == 0 && activeTeam.active != 1 && activeTeam.mons[1].hp > 0 && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeTeam.switchCooldown == 0 && activeTeam.active != 1 && activeTeam.mons[1].hp > 0 && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeTeam.switchCooldown == 0 && activeTeam.active != 2 && activeTeam.mons[2].hp > 0 && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeTeam.switchCooldown == 0 && activeTeam.active != 2 && activeTeam.mons[2].hp > 0 && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeTeam.switchCooldown == 0 && activeTeam.active != 3 && activeTeam.mons[3].hp > 0 && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeTeam.switchCooldown == 0 && activeTeam.active != 3 && activeTeam.mons[3].hp > 0 && activeMon.hp > 0) ? 1.0 : 0.0,
-                                (activeMon.hp < 0 && activeTeam.mons[1].hp > 0) ? 1.0 : 0.0,
-                                (activeMon.hp < 0 && activeTeam.mons[1].hp > 0) ? 1.0 : 0.0,
-                                (activeMon.hp < 0 && activeTeam.mons[2].hp > 0) ? 1.0 : 0.0,
-                                (activeMon.hp < 0 && activeTeam.mons[2].hp > 0) ? 1.0 : 0.0,
-                                (activeMon.hp < 0 && activeTeam.mons[3].hp > 0) ? 1.0 : 0.0,
-                                (activeMon.hp < 0 && activeTeam.mons[3].hp > 0) ? 1.0 : 0.0]
+                                (activeMon.energy >= activeMon.chargedMoves[1].energy && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.energy >= activeMon.chargedMoves[1].energy && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.energy >= activeMon.chargedMoves[2].energy && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.energy >= activeMon.chargedMoves[2].energy && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeTeam.switchCooldown == 0 && activeTeam.active != 1 && activeTeam.mons[1].hp > Int16(0) && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeTeam.switchCooldown == 0 && activeTeam.active != 1 && activeTeam.mons[1].hp > Int16(0) && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeTeam.switchCooldown == 0 && activeTeam.active != 2 && activeTeam.mons[2].hp > Int16(0) && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeTeam.switchCooldown == 0 && activeTeam.active != 2 && activeTeam.mons[2].hp > Int16(0) && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeTeam.switchCooldown == 0 && activeTeam.active != 3 && activeTeam.mons[3].hp > Int16(0) && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeTeam.switchCooldown == 0 && activeTeam.active != 3 && activeTeam.mons[3].hp > Int16(0) && activeMon.hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.hp < Int16(0) && activeTeam.mons[1].hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.hp < Int16(0) && activeTeam.mons[1].hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.hp < Int16(0) && activeTeam.mons[2].hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.hp < Int16(0) && activeTeam.mons[2].hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.hp < Int16(0) && activeTeam.mons[3].hp > Int16(0)) ? 1.0 : 0.0,
+                                (activeMon.hp < Int16(0) && activeTeam.mons[3].hp > Int16(0)) ? 1.0 : 0.0]
 end
 
 function play_decision(state::BattleState, decision::Int64)
